feat(chat): unified business_registry_lookup tool + shared dispatch module
Per-registry chat tools (ares_lookup_company, ares_search_companies)
invented a separate AI surface for each jurisdiction. Now that
contacts has a normalised cross-registry REST endpoint, the chat
tool should reuse the same shape: one tool, jurisdiction-aware enum,
identical dispatch path.

## Shared dispatch (apps/api/src/lib/business-registries/dispatch.ts)

Extracts BUSINESS_REGISTRY_DISPATCH, RegistryHandler, and the
normaliser helpers out of the contacts handler. Adds a `country:
CountryCode` field to each RegistryHandler so the chat layer can
route a user-facing jurisdiction (CZ, NO) to the right adapter.
Exports `executeRegistryLookup({ handler, query })` which runs the
shape-check → lookup-or-search → normalise → map-errors flow. Both
the REST endpoint and the chat tool now drive their lookups through
this single helper, so the two surfaces never drift in error mapping
or shape detection.

The REST handler at apps/api/src/handlers/contacts/business-registries-lookup.ts
slims down to ~60 lines: query schema, native-tool gating, hand off.

## New chat tool (apps/api/src/handlers/chat/tools/business-registry-tools.ts)

`createBusinessRegistryTools({ enabledJurisdictions })` registers
`business_registry_lookup` with a dynamic `jurisdiction` picklist
derived from `enabledJurisdictions`. The list is empty → the tool
isn't registered at all (no dead picker for the model). The
description points the model at "use jurisdiction for the country
whose registry should be searched, for example CZ for the Czech
ARES register or NO for Brønnøysundregistrene in Norway." The
execute function looks up the handler by country and runs the
shared dispatch.

## Wiring (apps/api/src/handlers/chat/tools/chat-tools.ts)

- Replaces createAresTools with createBusinessRegistryTools.
- Computes enabledJurisdictions from BUSINESS_REGISTRY_DISPATCH:
  every shipped adapter whose nativeToolSlug is not in
  disabledNativeToolSlugs contributes its country to the enum.
- BUILT_IN_CHAT_TOOL_POLICY_KINDS swaps the two ares_* entries for
  one business_registry_lookup entry, classified publicOfficial.
- ares-tools.ts is deleted; the chat surface shrinks accordingly.

## Toggleable enablement (packages/catalogue/src/loader.ts)

Adds `brreg` to TOGGLEABLE_NATIVE_TOOL_BACKEND_SLUGS. The recommended
mapping NO → brreg already exists; this finally surfaces the tool
in the MCP settings page and in the chat tool's jurisdiction enum
when a Norwegian org is configured.

## Frontend (apps/web/src/components/chat/...)

- chat-ui-tools.ts: adds business_registry_lookup to CHAT_TOOL_TITLE_KEYS
  and PUBLIC_OFFICIAL_CHAT_TOOL_NAMES; moves ares_lookup_company and
  ares_search_companies to LEGACY_CHAT_TOOL_TITLE_KEYS so historical
  chat history keeps rendering with a recognisable label.
- tool-call-card.tsx: NATIVE_TOOL_BRANDS retains the two ares entries
  as historical aliases (comment makes that explicit).
- i18n: adds chat.tool.business_registry_lookup to en.json; the i18n
  codegen propagates the key to every locale + messages.gen.ts.

## Tests

- business-registry-tools.test.ts: empty enabledJurisdictions →
  tool not registered; one/two jurisdictions → tool registered.
- catalog-metadata.test.ts: brreg now expected in the toggleable
  catalogue (was previously asserted absent).
- tool-schema.test.ts: chat tools test swaps ares_lookup_company /
  ares_search_companies for business_registry_lookup; policy assertion
  unchanged (publicOfficial).
- shared/errors.test.ts: follow-up tests for the EntityNotFound
  dual-detection contract that missed PR #554's cutoff (per-adapter
  instanceof + shared isEntityNotFound both work). Locks in the
  pattern so a future refactor that breaks either path fails CI.
