+++ apps/web/src/i18n/langs/cs.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "Struktura normy BOE",
       "boe_search_legislation": "Vyhledávání v BOE",
       "borme_get_summary": "Denní souhrn BORME",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Připravuji dokument",
       "describe-stella-api": "Připravuji vyhledání dat",
       "describe-stella-function": "Připravuji vyhledání dat",
