+++ apps/web/src/i18n/langs/en.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "BOE law structure",
       "boe_search_legislation": "BOE legislation search",
       "borme_get_summary": "BORME daily summary",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Preparing document",
       "describe-stella-api": "Preparing data lookup",
       "describe-stella-function": "Preparing data lookup",
