+++ apps/web/src/i18n/langs/et.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "BOE normi struktuur",
       "boe_search_legislation": "BOE seaduste otsing",
       "borme_get_summary": "BORME päevane kokkuvõte",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Dokumendi ettevalmistamine",
       "describe-stella-api": "Andmepäringu ettevalmistamine",
       "describe-stella-function": "Andmepäringu ettevalmistamine",
