+++ apps/web/src/i18n/langs/fr.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "Structure d’une norme BOE",
       "boe_search_legislation": "Recherche dans le BOE",
       "borme_get_summary": "Sommaire quotidien BORME",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Préparation du document",
       "describe-stella-api": "Préparation de la recherche de données",
       "describe-stella-function": "Préparation de la recherche de données",
