+++ apps/web/src/i18n/langs/lt.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "BOE normos struktūra",
       "boe_search_legislation": "BOE teisės aktų paieška",
       "borme_get_summary": "BORME dienos suvestinė",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Ruošiamas dokumentas",
       "describe-stella-api": "Ruošiama duomenų paieška",
       "describe-stella-function": "Ruošiama duomenų paieška",
