+++ apps/web/src/i18n/langs/lv.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "BOE normas struktūra",
       "boe_search_legislation": "BOE tiesību aktu meklēšana",
       "borme_get_summary": "BORME dienas kopsavilkums",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Dokumenta sagatavošana",
       "describe-stella-api": "Datu meklēšanas sagatavošana",
       "describe-stella-function": "Datu meklēšanas sagatavošana",
