+++ apps/web/src/i18n/langs/pl.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "Struktura aktu BOE",
       "boe_search_legislation": "Wyszukiwanie w BOE",
       "borme_get_summary": "Dzienne podsumowanie BORME",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Przygotowywanie dokumentu",
       "describe-stella-api": "Przygotowywanie odczytu danych",
       "describe-stella-function": "Przygotowywanie odczytu danych",
