+++ apps/web/src/i18n/langs/sk.json
@@ -510,6 +510,7 @@
       "boe_get_law_structure": "Štruktúra normy BOE",
       "boe_search_legislation": "Vyhľadávanie v BOE",
       "borme_get_summary": "Denný súhrn BORME",
+      "business_registry_lookup": "Searching business registry",
       "create-document": "Pripravujem dokument",
       "describe-stella-api": "Pripravujem vyhľadanie dát",
       "describe-stella-function": "Pripravujem vyhľadanie dát",
